Migrate to configuring the KubernetesClient via KUBECONFIG
We were previously passing in a master URL to point the scheduler to a specific Kubernetes
cluster. This seems to be an anti-pattern in the Kubernetes world. Instead, we allow
the client to pick up the configuration via the KUBECONFIG variable in the environment.
This significantly simplifies testing.

In the process of doing so, we've also removed the "CoreV1Api" kubernetes-java client,
and now only use the fabric8.io client.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>