Add the or-tools CP-SAT backend (#9)
Adds the or-tools CP-SAT backend.

This is a large feature addition that introduces an or-tools CP-SAT backend. It accepts as input the SQL views expressed as comprehensions and generates a Java program that encodes the corresponding model using the Java bindings for the or-tools CP-SAT solver. For now, the Model internally compiles and runs the generated Java program, but in the future, we'll make it so that the generated program can simply be added to a project's source tree in the spirit of protobufs.

In addition, this commit adds some changes on the comprehension AST, and accommodates those changes on the Minizinc side as well. It also adds parameterized testing using both solvers for existing tests (like the ModelTest).

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>