Compiler improvements: introduce intermediate variables, an output IR, and use scalar products instead of sum decompositions where possible (#43)
The first iteration of the or-tools compiler generated nested for loops for views, and computed in a single Java statement, any expressions that were extracted from these loops (constraints or result-sets). This strategy came in the way of performing many natural optimizations for reducing the load on the solver, such as re-using variables across expressions, and avoiding unnecessary decompositions of constraints (which in turn again creates too many variables).

This PR incorporates a set of dependent changes to address the above issue.

First, it introduces intermediate variables that hold the result of expressions. To fully pull this off however, especially for aggregates, it required us to extract lists of intermediate variables from nested for loops that we may have already generated. This previous point was a symptom of the fact that we were directly generating Java code as we parse the comprehension AST, rather than generating an IR.

Our second change is therefore the addition of an output IR that we can modify as we learn new expressions that are being computed across all views. The output IR is currently rudimentary, but gives us a base for evolving the compiler.

Third, the above changes including the output IR, allows us to perform smarter optimizations for some kind of expressions: one example of this is to use scalar products whenever we compute sums of expressions of the form X * Y, where X is a variable expression and Y is a constant.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>